Rename client_consensus.rs to consensus_state.rs
