+++ modules/src/core/ics02_client/mod.rs
@@ -1,9 +1,9 @@
 //! ICS 02: Client implementation for verifying remote IBC-enabled chains.
 
-pub mod client_consensus;
 pub mod client_def;
 pub mod client_state;
 pub mod client_type;
+pub mod consensus_state;
 pub mod context;
 pub mod error;
 pub mod events;
