+++ modules/src/mock/host.rs
@@ -9,8 +9,8 @@ use tendermint_testgen::light_block::TmLightBlock;
 use tendermint_testgen::{Generator, LightBlock as TestgenLightBlock};
 
 use crate::clients::ics07_tendermint::consensus_state::ConsensusState as TMConsensusState;
-use crate::core::ics02_client::client_consensus::ConsensusState;
 use crate::core::ics02_client::client_type::ClientType;
+use crate::core::ics02_client::consensus_state::ConsensusState;
 use crate::core::ics02_client::error::Error;
 use crate::core::ics02_client::header::{Header, TENDERMINT_HEADER_TYPE_URL};
 use crate::core::ics24_host::identifier::ChainId;
