+++ relayer/src/chain/handle.rs
@@ -29,9 +29,9 @@ use ibc::{
 
 use crate::{
     account::Balance,
-    client_consensus::{AnyConsensusState, AnyConsensusStateWithHeight},
     config::ChainConfig,
     connection::ConnectionMsgType,
+    consensus_state::{AnyConsensusState, AnyConsensusStateWithHeight},
     denom::DenomTrace,
     error::Error,
     event::monitor::{EventBatch, Result as MonitorResult},
