+++ relayer/src/chain/handle/base.rs
@@ -40,9 +40,9 @@ use crate::{
         },
         tracking::TrackedMsgs,
     },
-    client_consensus::{AnyConsensusState, AnyConsensusStateWithHeight},
     config::ChainConfig,
     connection::ConnectionMsgType,
+    consensus_state::{AnyConsensusState, AnyConsensusStateWithHeight},
     denom::DenomTrace,
     error::Error,
     keyring::KeyEntry,
