+++ relayer/src/lib.rs
@@ -24,9 +24,9 @@ pub mod account;
 pub mod cache;
 pub mod chain;
 pub mod channel;
-pub mod client_consensus;
 pub mod config;
 pub mod connection;
+pub mod consensus_state;
 pub mod denom;
 pub mod error;
 pub mod event;
